+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 ### Changes
 * Lowered Cryo Aspect freeze duration
+* Air stealer now scales with attack speed and no longer with weapon % attack damage
 
 ### Fixes
 * Player attack cooldown not calculated correctly in hammer knockback